This entire plist business is messy, but by default the port is retrieved from settings as a 64-bit long long. Now, setting it later causes it to be cast back to uint16_t, which ruins all calls _after_ the call to link::init_link(), the only cure I see is to forcedly cast it back to long long when setting.

+++ lib/link.cpp
@@ -83,7 +83,7 @@ link_host_state::init_link(settings_provider* settings, uint16_t default_port)
     if (settings) {
         auto s_port = settings->get("port");
         if (!s_port.empty()) {
-            int port = boost::any_cast<int>(s_port);
+            int port = boost::any_cast<long long>(s_port);
             if (port > 0 && port <= 65535) {
                 default_port = port;
             }
@@ -133,7 +133,7 @@ link_host_state::init_link(settings_provider* settings, uint16_t default_port)
 
     // Remember the port number we ended up using.
     if (settings) {
-        settings->set("port", default_port);
+        settings->set("port", (long long)default_port);
     }
 }
